rename tag-steps and use their titles
