+++ packages/app/src/router/index.ts
@@ -38,12 +38,6 @@ const routes: RouteRecordRaw[] = [
     meta: { authentication: 'guests-only' },
   },
   {
-    path: '/tag/:tagId',
-    name: 'tag-details',
-    component: (): Component => import('../views/TagDetails.vue'),
-    props: true,
-  },
-  {
     path: '/sandbox',
     name: 'sandbox',
     component: (): Component => import('../views/Sandbox.vue'),
@@ -56,7 +50,13 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/tag/add',
     name: 'add-tag',
-    component: (): Component => import('../views/AddTag.vue'),
+    component: (): Component => import('../views/tag/AddTag.vue'),
+  },
+  {
+    path: '/tag/:tagId',
+    name: 'tag-details',
+    component: (): Component => import('../views/tag/TagDetails.vue'),
+    props: true,
   },
   {
     path: '/sandbox',
